requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-postgres>=0.0.12
psycopg[binary]>=3.1.0
flashrank>=0.2.0
streamlit>=1.30.0
pypdf>=4.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
